pyproject.toml
[tool.poetry]
name = "value-python"
version = "0.1.0"
description = ""
authors = ["srinivas <srinivas@valmi.io>"]
readme = "README.md"
packages = [{ include = "value", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
opentelemetry-api = ">=1.20.0"
opentelemetry-sdk = ">=1.20.0"
opentelemetry-exporter-otlp-proto-grpc = ">=1.20.0"
httpx = ">=0.24.0"
opentelemetry-instrumentation-langchain = { version = "^0.47.5", optional = true }
opentelemetry-instrumentation-google-generativeai = { version = "^0.47.5", optional = true }
opentelemetry-instrumentation = "^0.59b0"


[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
pytest-asyncio = ">=0.21.0"
pytest-cov = ">=4.1.0"
black = ">=23.0.0"
ruff = ">=0.1.0"
mypy = ">=1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
instrumentation = [
    "opentelemetry-instrumentation-langchain",
    "opentelemetry-instrumentation-google-generativeai",
]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
